+++ auto-update/test/Control/DebounceSpec.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE ScopedTypeVariables #-}
 module Control.DebounceSpec (spec) where
 
 import Control.Concurrent
@@ -88,7 +87,7 @@ getWaitAction = do
 -- | Get a debounce system with access to the internals for testing
 getDebounce :: DI.DebounceEdge -> IO (IORef Int, IO (), MVar (), IO ())
 getDebounce edge = do
-  ref :: IORef Int <- newIORef 0
+  ref <- newIORef 0
   let action = modifyIORef ref (+ 1)
 
   (waitAction, returnFromWait) <- getWaitAction
@@ -114,7 +113,7 @@ waitForBatonToBeTaken baton = waitUntil 5 $ tryReadMVar baton >>= (`shouldBe` No
 waitUntil :: Int -> IO a -> IO ()
 waitUntil n action = recovering policy [handler] (\_status -> void action)
   where policy = constantDelay 1000 `mappend` limitRetries (n * 1000) -- 1ms * n * 1000 tries = n seconds
-        handler _status = Handler (\(HUnitFailure {}) -> return True)
+        handler _status = Handler (\HUnitFailure{} -> return True)
 
 main :: IO ()
 main = hspec spec