+++ warp/Network/Wai/Handler/Warp/Run.hs
@@ -57,13 +57,13 @@ socketConnection set s = do
 socketConnection _ s = do
 #endif
     bufferPool <- newBufferPool
-    writeBuf <- allocateBuffer bufferSize
-    let sendall = sendAll' s
+    writeBuffer <- createWriteBuffer bufferSize
+    writeBufferRef <- newIORef writeBuffer
     isH2 <- newIORef False -- HTTP/1.x
     return Connection {
         connSendMany = Sock.sendMany s
       , connSendAll = sendall
-      , connSendFile = sendFile s writeBuf bufferSize sendall
+      , connSendFile = sendfile writeBufferRef
 #if MIN_VERSION_network(3,1,1)
       , connClose = do
             h2 <- readIORef isH2
@@ -76,14 +76,19 @@ socketConnection _ s = do
 #else
       , connClose = close s
 #endif
-      , connFree = freeBuffer writeBuf
       , connRecv = receive s bufferPool
       , connRecvBuf = receiveBuf s
-      , connWriteBuffer = writeBuf
-      , connBufferSize = bufferSize
+      , connWriteBuffer = writeBufferRef
       , connHTTP2 = isH2
       }
   where
+    sendfile writeBufferRef fid offset len hook headers = do
+      writeBuffer <- readIORef writeBufferRef
+      sendFile s (bufBuffer writeBuffer) (bufSize writeBuffer) sendall
+        fid offset len hook headers
+
+    sendall = sendAll' s
+
     sendAll' sock bs = UnliftIO.handleJust
       (\ e -> if ioeGetErrorType e == ResourceVanished
         then Just ConnectionClosedByPeer
@@ -309,7 +314,9 @@ fork set mkConn addr app counter ii = settingsFork set $ \unmask ->
         -- fact that async exceptions are still masked.
         UnliftIO.bracket mkConn cleanUp (serve unmask)
   where
-    cleanUp (conn, _) = connClose conn `UnliftIO.finally` connFree conn
+    cleanUp (conn, _) = connClose conn `UnliftIO.finally` do
+                          writeBuffer <- readIORef $ connWriteBuffer conn
+                          bufFree writeBuffer
 
     -- We need to register a timeout handler for this thread, and
     -- cancel that handler as soon as we exit.